Extract OpenAI code into lib
